Add Display module and update html
Display module is unfinished and lacks the ability to add and edit tasks
and projects, but it can display them. The html has been updated with
dialog boxes for adding and editing tasks and projects that are shown
when the user clicks the add or edit buttons. The css has been updated
to style the dialog boxes and the display of tasks and projects.